+++ src/main/java/carRacing/level3/model/strategy/RandomNum.java
@@ -1,4 +1,4 @@
-package carRacing.level3.service.strategy;
+package carRacing.level3.model.strategy;
 
 import java.util.Random;
