+++ src/test/java/carRacing/level3/RandomNumStrategyTest.java
@@ -2,8 +2,8 @@ package carRacing.level3;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import carRacing.level3.service.strategy.RandomMovingStrategy;
-import carRacing.level3.service.strategy.RandomNum;
+import carRacing.level3.model.strategy.RandomMovingStrategy;
+import carRacing.level3.model.strategy.RandomNum;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
